# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Kartala Task Management API",
        "title": "Kartala Task Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of tasks for the authenticated user with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting field (e.g. title asc, created_at desc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (todo, in_progress, done)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by title (ILike search)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTasksPaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new task for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_tasks.CreateTaskParameter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific task by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing task by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_tasks.UpdateTaskParameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a task by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "email": "must be a valid email address"
                    }
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 400
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request parameters"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerDeleteResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 200
                        },
                        "message": {
                            "type": "string",
                            "example": "Data deleted successfully"
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerInternalErrorResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 500
                        },
                        "message": {
                            "type": "string",
                            "example": "Internal server error, please try again later"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 404
                        },
                        "message": {
                            "type": "string",
                            "example": "Resource not found"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTask": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-21T15:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Implementing Swagger in Gin"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "todo"
                },
                "title": {
                    "type": "string",
                    "example": "Learn Swagger"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-02-21T15:00:00Z"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTaskResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTask"
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 200
                        },
                        "message": {
                            "type": "string",
                            "example": "Data fetched successfully"
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTasksPaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerTask"
                    }
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 200
                        },
                        "message": {
                            "type": "string",
                            "example": "Data fetched successfully"
                        },
                        "pagination": {
                            "type": "object",
                            "properties": {
                                "has_next": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "has_previous": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "limit": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "page": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "total": {
                                    "type": "integer",
                                    "example": 100
                                },
                                "total_pages": {
                                    "type": "integer",
                                    "example": 10
                                }
                            }
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 401
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid or expired session"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "internal_tasks.CreateTaskParameter": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_tasks.UpdateTaskParameter": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}